+++ src/main/scala/scalismo/ui/model/properties/OpacityProperty.scala
@@ -1,14 +1,14 @@
 package scalismo.ui.model.properties
 
 object OpacityProperty {
-  val DefaultValue: Float = 1.0f
+  val DefaultValue: Double = 1.0
 }
 
-class OpacityProperty(initialValue: Float) extends NodeProperty[Float](initialValue) {
+class OpacityProperty(initialValue: Double) extends NodeProperty[Double](initialValue) {
   def this() = this(OpacityProperty.DefaultValue)
 
-  override protected def sanitize(possiblyNotSane: Float): Float = {
-    Math.max(0f, Math.min(1f, possiblyNotSane))
+  override protected def sanitize(possiblyNotSane: Double): Double = {
+    Math.max(0.0, Math.min(1.0, possiblyNotSane))
   }
 }
